chore: testing out ssr and hydrate module
