+++ packages/ontario-design-system-component-library/package.json
@@ -12,11 +12,32 @@
 		"dist/",
 		"loader/",
 		"src/",
+		"hydrate/",
 		"www/",
 		"LICENSE",
 		"package.json",
 		"readme.md"
 	],
+	"exports": {
+		".": {
+			"types": "./dist/types/index.d.ts",
+			"import": "./dist/index.mjs",
+			"require": "./dist/index.cjs.js",
+			"default": "./dist/index.mjs"
+		},
+		"./loader": {
+			"types": "./loader/index.d.ts",
+			"import": "./loader/index.js",
+			"require": "./loader/index.js",
+			"default": "./loader/index.js"
+		},
+		"./components": {
+			"types": "./dist/types/components.d.ts",
+			"import": "./dist/index.js",
+			"require": "./dist/index.cjs.js",
+			"default": "./dist/index.js"
+		}
+	},
 	"scripts": {
 		"prebuild": "pnpm run build:config",
 		"build": "pnpm run build:stencil && pnpm run format && pnpm run docs:copy:readme",
@@ -41,7 +62,7 @@
 	},
 	"dependencies": {
 		"@ongov/ontario-design-system-global-styles": "5.0.0-alpha.3",
-		"@stencil/core": "^4.23.0",
+		"@stencil/core": "4.30.0",
 		"dotenv": "^16.4.5",
 		"uuid": "^11.0.5"
 	},
